Move symbol index to subpackage
